+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Add customization parameter 'hideRightMenu' for hiding right panel on first loading (bug #39096)
 
 ### Document Editor
+* Search selected text
+* Add blank page
 
 ### Spreadsheet Editor
 * Set print area